server/requirements.txt
# core deps
python-dotenv
flask
flask-cors
flask-script
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
psycopg2-binary
geopy
gitpython
gunicorn
pandas
requests
# testing stuff
pytest
factory-boy
pytest-cov
faker
